# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Foodie - Cooking & Food Ordering Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      summary: Get current user's wishlist
      responses:
        "200":
          description: Wishlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add item to wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToWishlistBody"
      responses:
        "201":
          description: Added to wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistItem"
        "409":
          description: Already in wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /wishlist/{id}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove item from wishlist
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shorts:
    get:
      operationId: listShorts
      tags: [shorts]
      summary: List cooking shorts
      responses:
        "200":
          description: List of cooking shorts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Short"
    post:
      operationId: createShort
      tags: [shorts]
      summary: Create a cooking short (cafe only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShortBody"
      responses:
        "201":
          description: Short created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Short"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shorts/{id}:
    get:
      operationId: getShort
      tags: [shorts]
      summary: Get a cooking short
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Short
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Short"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteShort
      tags: [shorts]
      summary: Delete a cooking short
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Short deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shorts/{id}/like:
    patch:
      operationId: likeShort
      tags: [shorts]
      summary: Like a cooking short
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Short liked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Short"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    FoodItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        category:
          type: string
        cafeId:
          type: integer
        cafeName:
          type: ["string", "null"]
        available:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - price
        - imageUrl
        - category
        - cafeId
        - cafeName
        - available
        - rating
        - reviewCount
        - createdAt
...
    WishlistItem:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        foodItemId:
          type: integer
        foodItem:
          $ref: "#/components/schemas/FoodItem"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - foodItemId
        - foodItem
        - createdAt
    AddToWishlistBody:
      type: object
      properties:
        foodItemId:
          type: integer
      required:
        - foodItemId
    Short:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        videoUrl:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        authorId:
          type: integer
        authorName:
          type: string
        authorRole:
          type: string
        views:
          type: integer
        likes:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - videoUrl
        - thumbnailUrl
        - authorId
        - authorName
        - authorRole
        - views
        - likes
        - createdAt
    CreateShortBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
        thumbnailUrl:
          type: string
      required:
        - title
        - videoUrl
...